moved avionics images folder from images to avionics
