+++ frontend/src/components/Avionics/AvionicsCarousel.css
@@ -1,21 +1,21 @@
 #AvionicsHero_carouselImg1 {
-    /*background-image: url("src/images/bg-graphics/screen-avionics/avionics-slider/avionics2.png");*/
-    /*background-image: url("src/images/bg-graphics/home-screen/home-slider/fullscreen-slider2.png");*/
+    background-image: url("./avionics-slider/avionics2.png");
+    /*background-image: url("/avionics-slider/avionics2.png");*/
     background-position: 50% 10%;
     transform: rotateX(25deg);
 }
 
 #AvionicsHero_carouselImg2 {
-    /*background-image: url("src/images/bg-graphics/screen-avionics/avionics-slider/avionics2.png");*/
+    background-image: url("./avionics-slider/avionics2.png");
     background-position: 50% 10%;
 }
 
 #AvionicsHero_carouselImg3{
-    /*background-image: url("src/images/bg-graphics/screen-avionics/avionics-slider/avionics3.png");*/
+    background-image: url("./avionics-slider/avionics3.png");
     background-position: 50% 30%;
 }
 
 #AvionicsHero_carouselImg4{
-    /*background-image: url("src/images/bg-graphics/screen-avionics/avionics-slider/avionics4.png");*/
+    background-image: url("./avionics-slider/avionics4.png");
     background-position: 50% 30%;
 }
