+++ frontend/src/components/Avionics/AvionicsScreen.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import AvionicsCarousel from "./AvionicsCarousel.css"
+import AvionicsCarousel from "./AvionicsCarousel";
 import Footer from "../HeaderFooter/Footer";
 
 
 const AvionicsScreen = () => {
     return (
         <>
-            <AvionicsCarousel/>
+           <AvionicsCarousel/>
             <h1>hello world</h1>
 
             <Footer/>
