+++ frontend/src/components/Avionics/avionics-slider/avionics1.png
