+++ frontend/src/components/Avionics/avionics-slider/avionics2.png
