+++ frontend/src/components/Avionics/avionics-slider/avionics3.png
