+++ frontend/src/components/Avionics/avionics-slider/avionics4.png
